Mypyc demo
